+++ package.json
@@ -19,6 +19,8 @@
     "@mui/material": "^5.15.11",
     "@mui/x-data-grid": "^6.19.5",
     "@react-oauth/google": "^0.12.1",
+    "@stripe/react-stripe-js": "^2.7.0",
+    "@stripe/stripe-js": "^3.3.0",
     "flowbite-react": "^0.7.2",
     "leaflet": "^1.9.4",
     "react": "^18.2.0",
