+++ src/components/Navigationbar.jsx
@@ -27,7 +27,7 @@ const Navigationbar = () => {
                         Services
                     </li>
                 </RouterLink>
-                <RouterLink to="/form">
+                <RouterLink to="/Cart">
                     <li className='my-4 py-4 border-b border-slate-800 hover:bg-slate-800 hover:rounded'>
                         Place Order
                     </li>
@@ -50,7 +50,7 @@ const Navigationbar = () => {
                     <Link spy={true} smooth={true} duration={500} to="Home" offset={-70} className={navLinkClass}>Home</Link>
                     <Link spy={true} smooth={true} duration={500} to="About" offset={-70} className={navLinkClass}>About</Link>
                     <RouterLink to="/service" className={navLinkClass}>Services</RouterLink>
-                    <RouterLink to="/form" className={navLinkClass}>Place Order</RouterLink>
+                    <RouterLink to="/Cart" className={navLinkClass}>Place Order</RouterLink>
                     <Link spy={true} smooth={true} duration={500} to="Contact" offset={-70} className={navLinkClass}>Contact</Link>
                     <RouterLink to="/Login" className={navLinkClass}>Sign In</RouterLink>
                 </ul>
